VM_Arm: Use correct types

+++ apps/Arm/vm_introspect/src/init_dataport_ram.c
@@ -12,8 +12,8 @@
 #include <sel4vmmplatsupport/guest_memory_util.h>
 #include <vmlinux.h>
 
-extern unsigned long linux_ram_base;
-extern unsigned long linux_ram_size;
+extern uintptr_t linux_ram_base;
+extern size_t linux_ram_size;
 
 extern dataport_caps_handle_t memdev_handle;
